requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
django_filter==23.2
django-dotenv==1.4.2
dj-rest-auth==4.0.1
dj-rest-auth[with_social]==4.0.1
djangorestframework-simplejwt==5.2.2
drf-yasg
python-dotenv
wheel
measurement==3.2.2